move message_log_widget to game
