+++ cockatrice/src/client/tabs/tab_game.cpp
@@ -11,13 +11,13 @@
 #include "../../game/game.h"
 #include "../../game/game_scene.h"
 #include "../../game/game_view.h"
+#include "../../game/log/message_log_widget.h"
 #include "../../game/player/player.h"
 #include "../../game/player/player_list_widget.h"
 #include "../../game/replay.h"
 #include "../../game/zones/card_zone.h"
 #include "../../main.h"
 #include "../../server/abstract_client.h"
-#include "../../server/message_log_widget.h"
 #include "../../server/user/user_list_manager.h"
 #include "../../settings/cache_settings.h"
 #include "../network/replay_timeline_widget.h"
