+++ cockatrice/src/client/tabs/tab_game.h
@@ -3,8 +3,8 @@
 
 #include "../../client/tearoff_menu.h"
 #include "../../game/abstract_game.h"
+#include "../../game/log/message_log_widget.h"
 #include "../../game/player/player.h"
-#include "../../server/message_log_widget.h"
 #include "../replay_manager.h"
 #include "../ui/widgets/visual_deck_storage/visual_deck_storage_widget.h"
 #include "pb/event_leave.pb.h"
