+++ cockatrice/src/game/log/message_log_widget.cpp
@@ -1,13 +1,13 @@
 #include "message_log_widget.h"
 
-#include "../client/sound_engine.h"
-#include "../client/tabs/tab_game.h"
-#include "../client/translate_counter_name.h"
-#include "../game/board/card_item.h"
-#include "../game/phase.h"
-#include "../game/player/player.h"
-#include "../game/zones/card_zone.h"
-#include "../settings/card_counter_settings.h"
+#include "../../client/sound_engine.h"
+#include "../../client/tabs/tab_game.h"
+#include "../../client/translate_counter_name.h"
+#include "../../settings/card_counter_settings.h"
+#include "../board/card_item.h"
+#include "../phase.h"
+#include "../player/player.h"
+#include "../zones/card_zone.h"
 #include "pb/context_move_card.pb.h"
 #include "pb/context_mulligan.pb.h"
 #include "pb/serverinfo_user.pb.h"
