+++ cockatrice/src/game/log/message_log_widget.h
@@ -1,9 +1,9 @@
 #ifndef MESSAGELOGWIDGET_H
 #define MESSAGELOGWIDGET_H
 
-#include "../client/translation.h"
-#include "../game/zones/logic/card_zone_logic.h"
-#include "chat_view/chat_view.h"
+#include "../../client/translation.h"
+#include "../../server/chat_view/chat_view.h"
+#include "../zones/logic/card_zone_logic.h"
 #include "user_level.h"
 
 class AbstractGame;
